adjustments on entriesList
