+++ components/PictureDetails.js
@@ -3,13 +3,7 @@ import StyledImage from "./StyledImage";
 import { styled } from "styled-components";
 import { useRouter } from "next/router";
 
-export default function PictureDetails() {
-  // ----- nur das Bild, das auf OVerview/my Art angeklickt wurde, wird dargestellt
-  const router = useRouter();
-  const routerId = router.query.index; //  gibt IMMER einen String zurück!!!! id auch besser als String, da kein mathematischer Anwendungsfalls vorhanden
-
-  const selectedPicture = pictures.find((picture) => picture.id === routerId);
-
+export default function PictureDetails({ selectedPicture }) {
   if (!selectedPicture) {
     return <p>Loading...</p>;
   }
